# share authentication lookups