… staging/src/k8s.io/apiextensions-apiserver/go.sum
… staging/src/k8s.io/apiserver/go.sum
… staging/src/k8s.io/kube-aggregator/go.sum
… staging/src/k8s.io/legacy-cloud-providers/go.sum
… staging/src/k8s.io/sample-apiserver/go.sum
… vendor/BUILD
… vendor/golang.org/x/sync/singleflight/BUILD
… vendor/golang.org/x/sync/singleflight/singleflight.go
… vendor/modules.txt