Still need to benchmark.

This is a version of @lavalamp's change that is compatible with the expiring cache.

/kind bug
/sig auth

```release-note
NONE
```