@mikedanese 2019-11-14
Here's some benchmarks taken with this change rebased ontop of the expiring cache change. The arguments to the benchmark are `-test.benchmem -test.bench='CachedTokenAuthenticator/.*' -test.count=10 -test.run='^$' -test.timeout=1h`. Since the benchtime is 1s (the default), these benchmarks are very load heavy. See the comment here:

https://github.com/kubernetes/kubernetes/blob/27067540ffc97495e7a63f402b4a2fda01f40085/staging/src/k8s.io/apiserver/pkg/authentication/token/cache/cached_token_authenticator_test.go#L199-L202

I'm collecting a separate set of benchmarks with a benchtime of 30s. Preliminary results show that this helps significantly with stampeding heard, which will be exacerbated in real world environments where the penalty of a lookup is higher than what we are simulating in our tests:

https://github.com/kubernetes/kubernetes/blob/27067540ffc97495e7a63f402b4a2fda01f40085/staging/src/k8s.io/apiserver/pkg/authentication/token/cache/cached_token_authenticator_test.go#L283-L286

```
name                                                  old time/op     new time/op     delta
CachedTokenAuthenticator/tokens=100_threads=1-12         1.34µs ± 1%     1.31µs ± 1%    -2.19%  (p=0.000 n=9+8)
CachedTokenAuthenticator/tokens=100_threads=16-12        1.36µs ± 4%     1.33µs ± 7%      ~     (p=0.113 n=10+9)
CachedTokenAuthenticator/tokens=100_threads=256-12       4.95µs ±21%     1.08µs ± 5%   -78.25%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=500_threads=1-12         19.8µs ± 4%     13.7µs ± 8%   -30.61%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=500_threads=16-12        2.20µs ±24%     1.29µs ± 6%   -41.08%  (p=0.000 n=9+10)
CachedTokenAuthenticator/tokens=500_threads=256-12       21.3µs ± 8%      1.2µs ± 2%   -94.50%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=2500_threads=1-12        82.5µs ± 1%     82.7µs ± 1%      ~     (p=0.280 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=16-12       19.4µs ± 1%     13.2µs ± 1%   -31.92%  (p=0.000 n=9+10)
CachedTokenAuthenticator/tokens=2500_threads=256-12      37.7µs ± 5%      5.1µs ±10%   -86.58%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=1-12       91.8µs ± 0%     91.4µs ± 1%      ~     (p=0.065 n=9+10)
CachedTokenAuthenticator/tokens=12500_threads=16-12      27.7µs ± 0%     24.8µs ± 0%   -10.72%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=256-12     43.4µs ± 4%     22.0µs ± 1%   -49.39%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=1-12       93.4µs ± 0%     92.9µs ± 0%    -0.61%  (p=0.000 n=8+10)
CachedTokenAuthenticator/tokens=62500_threads=16-12      29.4µs ± 0%     26.6µs ± 0%    -9.25%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=256-12     45.3µs ± 1%     26.3µs ± 0%   -41.92%  (p=0.000 n=9+9)

name                                                  old lookups/op  new lookups/op  delta
CachedTokenAuthenticator/tokens=100_threads=1-12           0.00 ± 4%       0.00 ± 1%   -29.54%  (p=0.000 n=10+8)
CachedTokenAuthenticator/tokens=100_threads=16-12          0.02 ±10%       0.00 ± 7%   -84.35%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=100_threads=256-12         0.16 ±24%       0.00 ±11%   -98.57%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=500_threads=1-12           0.21 ± 5%       0.14 ± 8%   -35.87%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=500_threads=16-12          0.05 ± 6%       0.01 ± 6%   -75.36%  (p=0.000 n=8+10)
CachedTokenAuthenticator/tokens=500_threads=256-12         0.78 ± 8%       0.01 ±16%   -98.53%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=1-12          0.91 ± 0%       0.89 ± 0%    -2.22%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=16-12         0.72 ± 1%       0.48 ± 0%   -32.90%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=2500_threads=256-12        1.39 ± 5%       0.17 ± 5%   -87.72%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=1-12         1.00 ± 0%       0.98 ± 0%    -1.99%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=16-12        1.03 ± 0%       0.91 ± 0%   -11.54%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=256-12       1.60 ± 5%       0.79 ± 1%   -50.54%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=62500_threads=1-12         1.02 ± 0%       1.00 ± 0%    -2.23%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=16-12        1.09 ± 1%       0.98 ± 0%    -9.48%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=256-12       1.68 ± 2%       0.96 ± 0%   -43.01%  (p=0.000 n=9+9)

name                                                  old alloc/op    new alloc/op    delta
CachedTokenAuthenticator/tokens=100_threads=1-12           296B ± 0%       298B ± 0%    +0.68%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=100_threads=16-12          303B ± 0%       308B ± 0%    +1.65%  (p=0.000 n=10+7)
CachedTokenAuthenticator/tokens=100_threads=256-12         332B ± 7%       336B ± 2%      ~     (p=0.108 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=1-12           327B ± 0%       414B ± 2%   +26.75%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=500_threads=16-12          305B ± 0%       283B ± 0%    -7.11%  (p=0.000 n=9+10)
CachedTokenAuthenticator/tokens=500_threads=256-12         673B ± 9%       375B ± 1%   -44.25%  (p=0.000 n=10+8)
CachedTokenAuthenticator/tokens=2500_threads=1-12          435B ± 0%      1051B ± 0%  +141.59%  (p=0.000 n=6+10)
CachedTokenAuthenticator/tokens=2500_threads=16-12         418B ± 0%       711B ± 0%   +69.99%  (p=0.000 n=7+9)
CachedTokenAuthenticator/tokens=2500_threads=256-12        977B ±11%       678B ± 3%   -30.63%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=1-12         451B ± 0%      1131B ± 0%  +150.98%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=16-12        473B ± 0%      1111B ± 0%  +134.94%  (p=0.000 n=9+8)
CachedTokenAuthenticator/tokens=12500_threads=256-12     1.15kB ± 6%     1.50kB ± 1%   +30.42%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=1-12         453B ± 0%      1147B ± 0%  +153.15%  (p=0.000 n=8+9)
CachedTokenAuthenticator/tokens=62500_threads=16-12        483B ± 0%      1174B ± 0%  +142.81%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=256-12     1.20kB ± 6%     1.66kB ± 0%   +38.66%  (p=0.000 n=10+9)

name                                                  old allocs/op   new allocs/op   delta
CachedTokenAuthenticator/tokens=100_threads=1-12           11.0 ± 0%       11.0 ± 0%      ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=16-12          11.0 ± 0%       11.0 ± 0%      ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=256-12         11.0 ± 0%       12.0 ± 0%    +9.09%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=1-12           12.0 ± 0%       14.0 ± 0%   +16.67%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=16-12          11.6 ± 5%       11.0 ± 0%    -5.17%  (p=0.011 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=256-12         15.3 ± 5%       12.0 ± 0%   -21.57%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=1-12          16.0 ± 0%       27.0 ± 0%   +68.75%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=16-12         15.0 ± 0%       19.0 ± 0%   +26.67%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=256-12        18.7 ± 4%       17.0 ± 0%    -9.09%  (p=0.000 n=10+8)
CachedTokenAuthenticator/tokens=12500_threads=1-12         16.0 ± 0%       29.0 ± 0%   +81.25%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=16-12        16.0 ± 0%       28.0 ± 0%   +75.00%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=256-12       19.7 ± 4%       28.0 ± 0%   +42.13%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=1-12         16.0 ± 0%       29.0 ± 0%   +81.25%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=16-12        17.0 ± 0%       29.0 ± 0%   +70.59%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=256-12       20.0 ± 0%       30.0 ± 0%   +50.00%  (p=0.000 n=10+10)
```